Inventory design
Left Trigger spins inventory. On Start, list<> of sculptures are assigned to the inventory. I don't yet know how to OH WAIT maybe I make an enum with the number of possibilities equal to the number of sculptures. Then, when the selection point matches the range, the enum points the selector to the gameobject selected and spawn it for the player.